Added boolean to symbol table lookup to indicate identifier.
The main example where this might be a problem is where a function and a
TokenSym. In that case, it is plausible that either one may be the one in play,
but only one of them will need to have tcc_add_symbol() invoked on the compiler
state. For this reason, the extended symbol table lookup now gets a boolean
indicating if the symbol is being used as an identifier or not.

Note, the handling of structs may not be as clear cut as my current code
expects. That is, in the case of a struct, it may be necessary for the lookup
function to know if the struct is a type declaration or a global, and add the
symbol as appropriate. However, I believe that information can be obtained from
the TokenSym itself, so it is solvable by the callback function without any
additonal information.