Fixed symtab copy process for exsymtab functions.
The copied functions now corectly indicate that they are extern
functions, but they also retain the belief that they have already
been defined, i.e. that they are more than just a prototype. This
means that if a user tries to later redefine the function, the
compiler will correctly detect it and issue a redefinition error.