Suppress warning for null define symbol
The define symbol copy issues a warning if the tokensym does not have
a preprocessor slot. However, that should not be a warning if the
token index includes the SYM_FIELD flag. In that circumstance we have
a preprocessor argument, which should simply be a token, and has no
definition.