Retooled the extended symbol table approach.
This is a major revision to how the extended symbol table system works.
It is not debugged; it probably does not even compile. However, it
contains the heart of the ideas involved in revising the extended symbool
table system so that it can accomodate preprocessor redefinitions and
allow new entries in previously unused sym slots in a TokenSym. Both of
these very reasonable actions would have caused lots of trouble in the
previous extended symbol table implementation if one symbol table
depended on another and added something, and a third depended on the
first but not the second.

While extended symbol tables still copy the final symbol table at the
end of the compilation unit, consuming compilation units now copy
individual TokenSyms from their extended symbol tables rather than
using the original TokenSyms outright.

Next up: compile and test