Fix token id copying from extended to local symbol tables.
I had erroneously assumed that any token in a preprocessor macro
token stream had to refer to a token that was created in the
extended compilation context. This, of course, is not true. It
can include any allowed token, such as any non-identifier ASCII
symbol, like in this case, the parenthesis character.