Get test 42 to pass by with local stack and copy_extended_tokensym fixup
The machinery kicked off by copy_extended_tokensym makes quite a
bit of use of sym_push. The problem that test 42 revealed is that
the symbol that is needed happens to be called for the first time
in a block (which I expect will be the case most of the time),
then local_stack will be defined, and sym_push will push to that
location. Then, during the end-of-scope stack popping, the values
get reset! To fix this, I simply created a way to temporarily
turn the local stack, and later restore it.