Another substantial rewrite: simplified symtab creation and copy
I decided that I would like the relocation stage to automatically
add symbol pointers to the extended symbol table when tcc_relocate
was called. This meant that the extended symbol table would need to
be part of the TCCState struct. But, if it's part of the struct,
then I don't need a crazy callback during the compilation stage to
save it. I can simply create the extended symbol table and save the
pointer in the state. Done.

All of that led to what I think is a considerable simplification of
the extended symbol table copy API. Now, before compiling, you
simply call a function to tell tcc that this compiler state should
make a copy of the symtab before clearing up the stacks. If you do
not care about the symbol pointers (i.e. you don't want to or don't
plan to look up function pointers), then you can retrieve the
extended symbol table immediately after compilation. If you do want
function pointers, then you wait to retrieve the symbol table until
after relocation.