Add compiler prep stage
Extended symbols that refer to known tokens (such as "loop" or "file")
give trouble, as indicated in the test added by the previous commit.
This patch adds a rather large set of changes that make it possible to
add extended symbol table information for these tokens before the
compilation begins. Thus, it is called a compilation prep.

This commit changes the extended symbol table API (again), so I had
to change the test suite accordingly. Similarly, I will need to modify
C::Blocks to use the new API.