Differentiated define and other syms, but now get segfaults.
In the previous commit, test 80 failed because the serialization
code found a prev_tok field that unexpectedly had stuff. I decided
to rewrite things to keep the define and other Sym lists separate.
I like the new code, and it doesn't get caught on the previous
issue (indeed, that catching code was removed completely) but the
same Sym gives other issues: it has a weird token stream that
leads to segaults.