Added command-line capabilities for caching
I added two command-line switches. One serializes the extended
symbol table and the other writes the names of all identifiers
in an extended symbol table. Both of these only accept one file
name as an argument, so if you specify multiple input file names,
only the output for the last file will be saved to the file. This
should probably be extended some day.