Tweak exsymtab test 42; remove 45
Test 42 checks that function pointers have the same value, but I
found that on Windows they can have different values yet give the
correct behavior. If test 42 encounters a difference, it emits a
message explaining and continues.

Test 45 appears to test everything that test 42 covered, but for
only two contexts. This seemed odd, so I removed it.