Comment-out checks for exsymtab
A lot of work went in to adding exsymtab as a configuration option,
including guarding exsymtab-specific code with #ifdef blocks.
Unfortunately, these ifdef blocks are included in libtcc.h, which
complicates the use of exsymtab in consuming code. Furthermore,
win32 did not properly set up the preprocessor definitions.

If we want to make exsymtab a configuration option, then both *nix
and win32 build scripts should be smart enough to update config.h
(for cross-platform build process) and libtcc.h (for cross-platform
configurable API).