Update t/10 and t/70 for updated differing values of ret->r
Tests t/10 and t/70 check that function definitions and declarations are
the same with one exception, which is that the declaration is considered
a weak symbol. This is reflected in the associated register value,
ret->r. For some reason, the result of this has been altered, though I'm
really not clear how.

Fortunately (as indicated in the copious notes sprinkled in
tccexsymtab's get_new_symtab_pointer), the details of the contents don't
matter, its value is just copied to the exsymtab. This could cause
problems if these values change across cached symbol tables, but I'm not
going to tackle that just yet. (Presumably I could add a version
indicator to the beginning of the cache files and check against that...)