Demonstrate and solve the tokensym_list issue
Tests of C::Blocks on certain architectures were segfaulting. The
problem was traced to an overly trusting tcc_prep_tokensym_list.
That is, it trusted my knowledge of symbol definition too much,
and that knowledge turned out to be eroneous. The test demonstrates
the problem; the adjustment of adding symtab_tok_start fixes the
problem indicated by tests.