Fix testing of cached function attributes
With commit a4a20360 on Feb 13, 2017, grischka renamed one of the
Attribute struct's bitfields, flipping its meaning in the process.
Rather than compare all bits in the field by testing against the
r field of the Sym struct, this commit adds checks for each member
of the a field in the Sym struct, which is probably what it should
have done in the first place.